Allow talkgroup list to be ordered by usage
Passing a ?recent=1 to the /talkgroups/ url, the page will be ordered by the recent_usage and then last_transmission of the talkgroup.

To update the recent usage periodically run the update_talkgroup_usage command from manage.py. Every 5 to 10 minutes would be good if you have TALKGROUP_RECENT_LENGTH at the default of 15

New setting added TALKGROUP_RECENT_LENGTH this is the number of minutes that is looked at to get the recent usage. Default is 15.

The recent usage is the sum of the play_length for all transmission on the talkgroup for the time in settings.TALKGROUP_RECENT_LENGTH when update_talkgroup_usage was last run

CHANGE_TYPE: Database